[ci] Move Linux tests to separate stage (#6593)
Fixes: https://github.com/xamarin/xamarin-android/issues/5701

Adds a Linux test stage to exercise the `Microsoft.Android.Sdk.Linux`
.nupkg produced by the Linux build.  This stage contains one job that
runs the `SmokeTests` category of our MSBuild tests.

A couple of updates have been made to `Xamarin.ProjectTools` to support
test environments that do not have a system-wide installation of classic
Xamarin.Android.  This effort was focused on getting the `SmokeTests`
category to run, however more test harness changes will be needed in 
in this area in the future.