[Xamarin.Android.Build.Tasks] Add Support for bundletool meta-data
Fixes: https://github.com/xamarin/xamarin-android/issues/5378

Context: 2dab2bc3725a3ca3d56c6767c60f0eecf6426f46

When uploading an `.apk` or `.aab` file which contains native
libraries such as AOT'd .NET assemblies to the Google Play store,
the Google Play Console reports the following warning:

> This App Bundle contains native code, and you've not uploaded
> debug symbols.  We recommend that you upload a symbol file to make
> your crashes and ANRs easier to analyse and debug.

Commit 2dab2bc3 addresses a similar warning regarding obfuscated
Java/Kotlin code.

Improve support for this scenario by adding a new
`$(AndroidCreateProguardMappingFile)` MSBuild property which controls
the creation of the ProGuard mapping file.

When True (the default when r8 is used), a ProGuard mapping file will
be created and added to `.apk`/`.aab` build outputs.  This increases
the size of resulting applications by ~1% for our unit tests, and
will remove the above warning for Google Play Store submissions.

To make `$(AndroidCreateProguardMappingFile)`=True *work*, add
support for a new `@(AndroidAppBundleMetaDataFile)` item group.
`%(AndroidAppBundleMetaDataFile.Include)` values are those used by
the [`bundletool build-bundle --metadata-file`][0] option:

> `--metadata-file=target-bundle-path:local-file-path`
>
>   * `target-bundle-path` specifies a path relative to the root of
>      the app bundle where you would like the metadata file to be
>      packaged, and
>   * `local-file-path` specifies the path to the local metadata file
>     itself.

When `$(AndroidCreateProguardMappingFile)`=True, we automatically add:

	<AndroidAppBundleMetaDataFile Include="com.android.tools.build.obfuscation/proguard.map:$(AndroidProguardMappingFile" />

which causes the ProGuard mapping file to be included into the bundle.
This only applies to `.aab` package formats.

For `.apk` files the ProGuard mapping file will not be included in the
`.apk`,  as it doesn't not support it. However the file will be copied
to the  `$(OutputPath)` along with the `.apk`. It can be manually uploaded
to the Google Play Store from there.

[0]: https://developer.android.com/studio/build/building-cmdline#bundletool-build