[tests] update MultiDexR8ConfigWithNoCodeShrinking
78dd0216 introduced the `MultiDexR8ConfigWithNoCodeShrinking(false)`
test that asserts we get the warning:

    R8 : warning : Missing class: java.lang.ClassValue

The latest version of r8 no longer emits this warning, so Google must
have *fixed* something here.

The test still seems useful to build a multi-dex application with a
`@(ProguardConfiguration)` item, so let's drop one test case and just
assert that the warning isn't there. The test is no longer
parameterized.