[Xamarin.Android.Build.Tasks] remove Microsoft.Android.Sdk.PropertyPages.targets (#6622)
This reverts e88b2b8.

The property pages are no longer used by the XamarinVS extension and
the import and `Microsoft.Android.Sdk.PropertyPages.targets.targets` file
unnecessary.

The file was also not included in `create-installers.targets`
(probably an oversight!). This means we never actually shipped this
file in any installers or NuGet packages.