[ci] Sign and publish Microsoft.Android.Sdk.Linux (#6613)
Fixes: https://github.com/xamarin/xamarin-android/issues/6472

A `Sign Linux Archive` job has been added to the `Prepare .NET Release`
stage to conditionally Authenticode sign the
`Microsoft.Android.Sdk.Linux` package produced by the Linux build.
The "Real" signing condition that was previously set on each job in
this stage has been moved up to the stage level for simplicity.

The signed version of `Microsoft.Android.Sdk.Linux` will now be pushed
to the dotnet6 feed and to Maestro on all official builds.  It will
also be pushed to NuGet.org when that manually triggered stage runs.

Note that the Linux package is *not officially supported*, and has
not received extensive testing.