[build] Remove aapt2 from bundletool during build
We've been removing the aapt2 executables included in bundletool.jar
during our CI builds on macOS only.  This is being done to simplify the
work needed for installer signing and notarization.  We should move this
logic to the project that builds bundletool.jar to establish consistent
behavior across build environments.