[Xamarin.Android.Build.Tasks] default $(AndroidUseAssemblyStore) to false for debug builds (#6660)
Context: https://github.com/xamarin/QualityAssurance/tree/master/Manual/CrossPlatformDebugging
Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1430409
Fixes: https://github.com/xamarin/xamarin-android/issues/6480

A `CrossPlatformDebugging` sample project in the QA repo fails to
debug with:

    System.InvalidProgramException
    Message=Invalid IL code in Mono.SystemDependencyProvider:Initialize (): IL_002e: endfinally

The problem goes away if you do one of:

* Turn *on* Fast Deployment
* Add `<AndroidUseAssemblyStore>false</AndroidUseAssemblyStore>`
* Change `$(DebugType)` to `portable`. (the project has `full`)

The problem appears to be:

1. `$(AndroidUseAssemblyStore)` defaults to `true` when you turn Fast
   Deployment *off*.

2. Assembly stores appear to have some issue when `DebugType=full` --
   that is where the crash occurs.

In this case let's address No. 1, as we don't *really* want the
assembly store to be used when debugging. I reordered default values
for MSBuild properties, so that `$(AndroidIncludeDebugSymbols)` is
taken into account.

I added parameters for a debugging test for `DebugType=full`, and I
updated another test that used to set `AndroidUseAssemblyStore=false`
that should not be needed.