[Xamarin.Android.Build.Tasks] Fix the `RemoveResourceDesignerStep` to work under .NET 6
The current implementation of the `RemoveResourceDesignerStep` does not
work under .NET 6. This is mostly down to the fact that we didn't
`override` the require methods to make it work in that environment.

This PR does also rework the `RemoveResourceDesignerStep` to split some
of the functionality out into a new base class. This new base class will
be used in the future by the Resource Assembly linker step. Since much of
the code is identical.