[Xamarin.Android.Build.Tasks] add libaot-*.dll.so files to dso_cache (#6695)
Context: c227042b6e58565e21dcd5e4ff5ea20fc4e9367c

In testing Profiled AOT, I found a problem in "Classic" Xamarin.Android
that was working in .NET 6:

 1. Enable AOT-related logging via:

        adb shell setprop debug.mono.log=all,mono_log_level=debug,mono_log_mask=aot

 2. Build & run a Release build using AOT

 3. Watch `adb logcat` output

I noticed that during (3) we are missing messages indicating that AOT
is being used! I was not able to find messages containing `libaot-`
or `AOT: FOUND method`.

Commit c227042b worked appropriately for .NET 6, because the build
ordering is:

 1. linker
 2. AOT
 3. `<GeneratePackageManagerJava/>`

The problem in "Classic" Xamarin.Android is that the ordering differs:

 1. linker
 2. `<GeneratePackageManagerJava/>`
 3. AOT

And so the list of native libraries would not contain any
`libaot-*.dll.so` files.

The difference of ordering, is due to how .NET 6 runs an "inner" build
per `$(RuntimeIdentifier)`.  It was much simpler to run the AOT
compiler inside the "inner" build to get things working.

To solve the problem on Classic, I think we should simply emit
`libaot-*.dll.so` file names in `<GeneratePackageManagerJava/>`.
Trying to make c227042b a .NET 6-only feature or reordering the build
seems like it would complicate things.

I updated a test for this scenario.

After this change, I can see the AOT log messages, and things are
working again:

	D Mono    : Prepared to set up assembly 'HelloWorld' (HelloWorld.dll)
	V Mono    : Loading assembly HelloWorld (0x741faf83e0) into domain RootDomain (0x749fadedf0) and ALC 0x0
	D Mono    : Assembly HelloWorld[0x741faf83e0] added to domain RootDomain, ref_count=1
	D monodroid-assembly: monodroid_dlopen: hash for name 'HelloWorld.dll.so' is 0xdacd7936300d6334
	D monodroid-assembly: dso_cache: looking for hash 0xdacd7936300d6334
	D monodroid-assembly: dso_cache: entry_hash == 0x8da9f07ad98f854a
	D monodroid-assembly: dso_cache: entry_hash == 0xd32c30d21c7615fc
	D monodroid-assembly: dso_cache: entry_hash == 0xe99c30c1484d7f4f
	D monodroid-assembly: dso_cache: entry_hash == 0xe156906e01e8a274
	D monodroid-assembly: dso_cache: entry_hash == 0xdacd7936300d6334
	D monodroid-assembly: monodroid_dlopen: hash match found, DSO name is 'libaot-HelloWorld.dll.so'
	I monodroid-assembly: Trying to load shared library '/data/app/~~-fEdInfI9lIjBKycz2PKJw==/com.xamarin.android.helloworld-Z3O_lA5hXtmCqwFkb8DAuw==/base.apk!/lib/arm64-v8a/libaot-HelloWorld.dll.so'
	W linker  : Warning: "/data/app/~~-fEdInfI9lIjBKycz2PKJw==/com.xamarin.android.helloworld-Z3O_lA5hXtmCqwFkb8DAuw==/base.apk!/lib/arm64-v8a/libaot-HelloWorld.dll.so" has no DT_SONAME (will use libaot-HelloWorld.dll.so instead) and will not work when the app moves to API level 23 or later (https://android.googlesource.com/platform/bionic/+/master/android-changes-for-ndk-developers.md#missing-soname-enforced-for-api-level-23) (allowing for now because this app's target API level is still 21)
	V Mono    : AOT: module HelloWorld.dll.so wants to load image 0: HelloWorld
	...
	D Mono    : AOT: FOUND method HelloWorld.MainActivity:OnCreate (Android.OS.Bundle) [0x7320dd3430 - 0x7320dd3630 0x7320dd3d21]