Default VerifyDependencyInjectionOpenGenericServiceTrimmability
This feature switch ensures that DynamicallyAccessedMembers are applied correctly to open generic types used in Dependency Injection.

In the base SDK, this switch is enabled when PublishTrimmed=true. However, Android apps don't set PublishTrimmed=true in Debug builds. So devs are missing out on this validation.

Conversely, in published apps, we don't want to pay the cost of doing this validation. It was showing up in startup profiles. So turning the feature switch off in Release builds.