[One .NET] fix incremental AOT builds
Fixes: https://github.com/xamarin/xamarin-android/issues/6732

A problem occurs when you do:

    > dotnet new android
    > dotnet build -c Release -r android-arm64
    > dotnet build -c Release -r android-arm64

After the second build with no changes, MSBuild targets run that
shouldn't -- and you end up with an `.apk` that is missing all the
AOT'd assemblies!

What is happening is the `Inputs` to the `_RemoveRegisterAttribute`
have the wrong path of a file:

    Input file "obj\Release\net6.0-android\android-arm64\android-arm64\link.flag" does not exist.

Removing `-r android-arm64` from the above commands solve the problem!

The issue here being:

1. We have an "outer" build where normally `$(RuntimeIdentifiers)`
   (plural) triggers an "inner" build for each RID.

2. Each inner build runs ILLink & AOT compilation.

3. The outer build collects the resulting files and does other work.

At step 3, we were using the wrong path for `$(_AndroidLinkFlag)` --
but only when `$(RuntimeIdentifier)` (singular) is used.

I fixed this problem, and updated various tests to catch this in the
future.

I would ideally love for this inner/outer build situation to be
handled by the dotnet/sdk, but this is what we have for now.