[One .NET] default $(TargetPlatformVersion) to 32
Currently, if you build with `net6.0-android` and leave off the
number, it defaults to API 31. Let's bump this to API 32, Android 12L.

I also fixed the `Inputs` of the `_GenerateSupportedPlatforms` target,
so it will rerun when `Configuration.props` changes.

We will likely need to coordinate the API level installed by various
Visual Studio installers.