[Mono.Android] JNIEnv.FindClass() can use StaticMethods.CallStaticObjectMethod()
Code paths using `JNIEnv.CallStaticObjectMethod` are using the "old"
way of doing Java.Interop. One such case is found when profiling the
.NET Podcast app:

    15.09ms Mono.Android!Android.Runtime.JNIEnv.FindClass
     6.25ms Mono.Android!Android.Runtime.JNIEnv.CallStaticObjectMethod

This is called in one place by:

    6.16ms Mono.Android!Android.Content.Intent..ctor(Android.Content.Context,System.Type)

If we instead define on `Java.Lang.Class`:

    internal static JniPeerMembers Members => _members;

We could access `Java.Lang.Class.Members.StaticMethods`, however in
this case we get the method ID from:

    internal static unsafe void Initialize (JnienvInitializeArgs* args)
    {
        // ...
        mid_Class_forName = args->Class_forName;

If we use the `IntPtr` here to create:

    new JniMethodInfo (args->Class_forName, isStatic: true)

We can instead use:

    JniEnvironment.StaticMethods.CallStaticObjectMethod (Java.Lang.Class.Members.JniPeerType.PeerReference, mid_Class_forName, parameters)

With this change in place, I'm not really seen a noticeable
performance difference, but it should help a small amount by
allocating less?