[docs] improve instructions for obtaining methods.txt
When recording JIT times, we didn't have instructions to set
/manifest/application/@android:debuggable="true". This is a crucial
step for Release builds.

I also noted that `counters.txt` is not useful in .NET 6. It is
mostly empty text such as:

    ## Runtime.register: type=Microsoft.NetConf2021.Maui.MainApplication, Microsoft.NetConf2021.Maui, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
    ## Runtime.register: type=Microsoft.Maui.MauiApplication, Microsoft.Maui, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
    ## Runtime.register: type=Microsoft.Maui.MauiApplication+ActivityLifecycleCallbacks, Microsoft.Maui

It might be worth removing support for this in .NET 6, but let's just
comment that it isn't useful for now.