[Mono.Android] avoid RegisterJniNatives() string allocation
Based on the comments:

https://github.com/xamarin/xamarin-android/pull/6708#discussion_r801077205

Viewing `dotnet trace` output for the .NET Podcast app, we actually see:

    122.71ms (0.26%) Mono.Android!Android.Runtime.JNIEnv.RegisterJniNatives(intptr,int,intptr,intptr,int)
      5.13ms (0.01%) System.Private.CoreLib!System.String.Ctor(char*,int,int)

Once per type that enters RegisterJniNatives(), we were allocating a
potentially large `System.String` that contains every Java method that a
C# type overrides. If we use `ReadonlySpan<char>`, we can avoid this
allocation for each call.

After this change, instead I see:

    1.23ms System.Private.CoreLib!System.String.Ctor(char*,int,int)

This call happens earlier in the method, which it doesn't look like we
can use `ReadonlySpan<char>` for:

    string typeName = new string ((char*) typeName_ptr, 0, typeName_len);
    var type = Type.GetType (typeName);

I suspect this change saves 3-4ms of startup time in the .NET Podcast
app on a Pixel 5 device.