Remove unnecessary background Task in AndroidMessageHandler
AndroidMessageHandler.SetupRequest returns a Task so overrides can be async. The base implementation checks the object hasn't been disposed. However, it is doing this check using Task.Run, which queues a work item to the ThreadPool. This is unnecessary and causes more threads to be spawned than what is needed.

The fix is to just check for disposed synchronously and return a completed Task.

Fix #6854