[Xamarin.Android.Build.Tasks] AndroidGenerateResourceDesigner & DTB (#6844)
Fixes: https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1440274

In some circumstances, when
[`$(AndroidGenerateResourceDesigner)`][0]=False and a custom
`Resource.Designer.cs` is provided in the project, loading a project
would cause Visual Studio IntelliSense to show an ambiguity with e.g.
`Resource.Layout.activity_main`, as the member exists in *both* the
custom `Resource.Designer.cs` *and* in the design-time-generated
`Resource.Designer.cs`.

This was because the design-time build was not taking
`$(AndroidGenerateResourceDesigner)` into consideration before
generating `Resource.designer.cs`.  Consequently, we would *always*
create a `Resource.designer.cs` file into the `designtime` directory.

Fix this by turning off the `designtime` build when
`$(AndroidGenerateResourceDesigner)`=False.

Additionally, update `Xamarin.Android.Legacy.targets` so that
`$(AndroidGenerateResourceDesigner)` is *always* set in Legacy builds,
with a default value of True.  This is to get around some issues under
legacy where `$(AndroidGenerateResourceDesigner)` was not defined.
This is because the `designtime` build tasks and targets are shared
between both Classic and .NET 6 build systems.

[0]: https://github.com/xamarin/xamarin-android/blob/4b33a6229c9babde795dea48d20ca86d8f9f9c97/Documentation/guides/building-apps/build-properties.md#androidgenerateresourcedesigner