[Mono.Android] Remove background Task in AndroidMessageHandler (#6857)
Fixes: https://github.com/xamarin/xamarin-android/issues/6854

`AndroidMessageHandler.SetupRequest()` returns a `Task` so that
method `override`s can be `async`.

The base implementation checks whether or not the object has been
disposed.  However, it is doing this check using `Task.Run()`, which
queues a work item to the `ThreadPool`.  This is unnecessary and
causes more threads to be spawned than is needed.

Instead of doing the disposed check as a Task, check synchronously
and return a completed task.