[Xamarin.Android.Build.Tasks] remove NuGet workaround for VS 2019
This reverts 7e9f595f.

We kept a workaround in place for VS 2019, as it didn't have a new
NuGet feature that was needed for .NET 6.

I believe we can remove this now, as .NET 6 requires VS 2022, anyway.
But we should be safe to remove it in xamarin-android/main now.