Implement marshal methods LLVM IR executable code generator.
The point of this commit is only to generate code and make sure it's
valid as far as compiling and linking are concerned. The code has not
been tested at run time as not all the infrastructure on the
Xamarin.Android side is implemented yet. This is on purpose, to keep PRs
smaller.

The majority of this PR introduces various classes, enums and structures
related to code generation. Support for various LLVM IR instructions is
limited only to those we actually use and only to elements of those
constructions that we use. As such, it's not a general purpose code
generator which allows us to make some assumptions and take some
shortcuts (without compromising correctness and validity of the
generated code)

Portions of the PR (the native type handling system) are to be treated
as proof-of-concept as they are not as optimized (design wise) as they
should be. The reason for this limitation is that it requires modifying
the previous LLVM IR data generation code and it would contribute to
this PR's already substantial size. The next PR in the series will take
care of that rewrite as well as it will focus on implementing the
runtime side of marshal methods, making it possible to actually run
applications which use marshal methods.

What this PR implements is the following:

  * LLVM IR
    * function and instruction attributes
    * function parameter (declaration/definition) and argument (runtime)
      handling
    * function variable (including parameters) handling, including
      unnamed local variables
    * support for native function signatures and pointers to functions
    * The ret, store, load, icmp, br, call and phi instructions
  * Marshal method generator
    * managed to JNI signature and symbol name translations