[marshal methods] Part 2 of ?
Implement a marshal method candidate classifier which is passed to JavaCallableWrapperGenerator
in order to identify methods that can be registered "statically" and for which we can generate
native code instead of using System.Reflection.Emit.

The selected methods are marked with the `[UnmanagedCallersOnly]` attribute and their support
code in the form of the connector method (which returns a delegate to the native callback) as
well as its backing field which stores the delegate, are removed from the assemblies.