Workaround MVIDs from dotnet/linker
Context: https://github.com/dotnet/runtime/issues/45649
Context: https://github.com/dotnet/linker/issues/2203

We are getting unique MVIDs per architecture output from the linker.

To workaround this problem, I used the "size" of the files instead. This works for now, but is not 100% correct. I think it would be fine for .NET 7 Android apps to use this for now in main. (hopefully temporary)