Fix LayoutBinding under AndroidX.
Fixes #6894

The `LayoutBinding.cs` class was using an incorrect namespace for
AndroidX. It was using `Androidx` rather than `AndroidX`. Also under
AndroidX the use of `FragmentManager` has been deprecated. Instead
we should be using `SupportFragmentManager` which is available on the
`AndroidX.Fragment.App.FragmentActivity` class.

Finally, remove the use of a `static` field to share object state between
tasks. We have been on `msbuild` rather than `xbuild` on *nix for a while
now, so this workaround is no longer needed.